+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ---
 
+#### 19.12.2018
+
+* [MainActivity] Adding to favourites onLongClick with a confirmation
+
 #### 18.12.2018
 
 * [Favourites] Removal confirmation dialog